backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
feedparser==6.0.11
beautifulsoup4==4.12.3
email-validator==2.1.1
fastapi-mail>=1.4.1
firebase-admin>=6.5.0
apscheduler>=3.10.4
